* Cleaning up the code

+++ oxts-decoder.cpp
@@ -70,10 +70,10 @@ std::pair<bool, std::pair<opendlv::proxy::GeodeticWgs84Reading, opendlv::proxy::
 
             // Normalize between -M_PI .. M_PI.
             while (northHeading < -M_PI) {
-                northHeading += 2.0f*M_PI;
+                northHeading += 2.0f*static_cast<float>(M_PI);
             }
             while (northHeading > M_PI) {
-                northHeading -= 2.0f*M_PI;
+                northHeading -= 2.0f*static_cast<float>(M_PI);
             }
 
             heading.northHeading(northHeading);
